add endpoint CA certificate handling
OpenStack clouds often have TLS termination with certs signed by
non-public CAs which cannot be verified with a standard set of CA certs
shipped in the core snap used by other snaps.

There is a way to provide a path to a ca-file in cloud.conf for k8s
in-tree OpenStack provider code which relies on gophercloud to use
that cert when creating an HTTP client for o7k endpoint communication.

The content of the CA certificate has to come from the integrator charm
which is not a subordinate, therefore, it must expose that cert over
a relation with other charms consuming data required to render
cloud.conf.